+++ apps/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @perceo/perceo
 
+## 0.4.2
+
+### Patch Changes
+
+- Change LLM to sonnet 4.5 for cost and better performance
+  - @perceo/observer-engine@3.0.2
+
 ## 0.4.1
 
 ### Patch Changes